[Docs] Document how tests work for contributing
It was difficult for me to understand what `phpunit` was used for in this project initially. This blurb explains that phpunit is just used to invoke `phpstan` on each individual file in the `tests/Features` directory. Hopefully it can be improved in the future as the test suites improve.l